added page func

+++ app/controllers/breweries_controller.rb
@@ -1,18 +1,20 @@
 class BreweriesController < ApplicationController
 
-	def index
-
-		Brewery.destroy_all
-
-		if params[:state]
-			Brewery.get_breweries_data('state',params[:state])
-		elsif params[:name]
-			Brewery.get_breweries_data('name',params[:name])
-		else
-			Brewery.get_breweries_data('state','District of Columbia')
+	def index	
+		if params[:page] == '0'
+			Brewery.destroy_all
 		end
 
-        render json: Brewery.all.sort_by{|brewery| brewery[:name]}
+		if params[:state] 
+			Brewery.get_breweries_data('state',params[:state], params[:page])
+		elsif params[:name]
+			Brewery.get_breweries_data('name', params[:name], '0')
+		else
+			Brewery.get_breweries_data('state','District of Columbia', '0')
+		end
+
+        render json: Brewery.all
+
     end
